requirements.txt
# Core ML and Deep Learning
tensorflow==2.15.0
numpy==1.26.2
pillow==10.1.0

# MLOps Tools
zenml==0.55.5
zenml-mlflow==0.1.2
mlflow==2.9.2
dvc[s3]==3.36.1
evidently==0.4.8

# Data Processing
pandas==2.1.3
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
boto3==1.34.0
minio==7.2.3
requests==2.31.0

# Additional dependencies
protobuf==4.25.1
werkzeug==3.0.1
gunicorn==21.2.0
